% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{bom_check_batchNo_overWrite}
\alias{bom_check_batchNo_overWrite}
\title{bom的批次回写规则}
\usage{
bom_check_batchNo_overWrite(batchNo = "PRD00000003")
}
\arguments{
\item{batchNo}{批号}
}
\value{
返回值
}
\description{
bom的批次回写规则
}
\examples{
bom_check_batchNo_overWrite()
}
