% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{bom_check_is_newVersion}
\alias{bom_check_is_newVersion}
\title{检验BOM是否新版本}
\usage{
bom_check_is_newVersion(
  conn = conn_vm_erp_test(),
  FBOMNumber = "BOM000001",
  FVersion = "001"
)
}
\arguments{
\item{conn}{连接}

\item{FBOMNumber}{BOM}

\item{FVersion}{版本号}
}
\value{

}
\description{
检验BOM是否新版本
}
\examples{
bom_check_is_newVersion()
}
