% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_util.R
\name{bom_getBillGroup}
\alias{bom_getBillGroup}
\title{BOM获取单据编码的组别}
\usage{
bom_getBillGroup(FItemNumber = "2.104.20.00026")
}
\arguments{
\item{FItemNumber}{物料编码}
}
\value{
返回值
}
\description{
BOM获取单据编码的组别
}
\examples{
bom_getBillGroup()
}
