% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_util.R
\name{bom_getBillGroupID}
\alias{bom_getBillGroupID}
\title{获取BOM单据编号的组别}
\usage{
bom_getBillGroupID(conn = conn_vm_erp_test(), FItemNumber = "2.104.20.00026")
}
\arguments{
\item{conn}{连接}

\item{FItemNumber}{单据编号}
}
\value{
返回组别内码
}
\description{
获取BOM单据编号的组别
}
\examples{
bom_getBillGroupID()
}
