% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{BOM_getBomSkip}
\alias{BOM_getBomSkip}
\title{根据单据编号查询是否BOM跳层}
\usage{
BOM_getBomSkip(conn = conn_vm_erp_test(), PMCode = "2.104.20.00034")
}
\arguments{
\item{conn}{连接}

\item{PMCode}{物料代码}
}
\value{
返回值
}
\description{
根据单据编号查询是否BOM跳层
}
\examples{
BOM_getBomSkip()
}
