% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{bom_getBomVersionNumber}
\alias{bom_getBomVersionNumber}
\title{获取BOM的产品版本信息}
\usage{
bom_getBomVersionNumber(conn = conn_vm_erp_test(), PMCode = "2.104.20.00034")
}
\arguments{
\item{conn}{连接}

\item{PMCode}{代码}
}
\value{
返回值
}
\description{
获取BOM的产品版本信息
}
\examples{
bom_getBomVersionNumber()
}
