% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{bom_getInterId}
\alias{bom_getInterId}
\title{获取BOM的内码}
\usage{
bom_getInterId(conn = conn_vm_erp_test(), PMCode = "2.104.20.00034")
}
\arguments{
\item{conn}{连接}

\item{PMCode}{物料代码}
}
\value{
返回BOM内码，为空则为0
}
\description{
获取BOM的内码
}
\examples{
bom_getInterId()
}
