% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{bom_getList}
\alias{bom_getList}
\title{获取待处理的BOM清单}
\usage{
bom_getList(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取待处理的BOM清单
}
\examples{
bom_getList()
}
