% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{BOM_getMaxBillValue}
\alias{BOM_getMaxBillValue}
\title{获取BOM单号流水号的最大值并转化为数值型}
\usage{
BOM_getMaxBillValue(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取BOM单号流水号的最大值并转化为数值型
}
\examples{
BOM_getMaxBillValue()
}
