% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_util.R
\name{BOM_getNewBillNo}
\alias{BOM_getNewBillNo}
\title{获取BOM的最新单据编号}
\usage{
BOM_getNewBillNo(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取BOM的最新单据编号
}
\examples{
BOM_getNewBillNo()
}
