% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{BOM_getNewBillTpl_Body}
\alias{BOM_getNewBillTpl_Body}
\title{获取BOM新增的模板_表体部分}
\usage{
BOM_getNewBillTpl_Body(conn = conn_vm_erp_test(), data_bom)
}
\arguments{
\item{conn}{连接}

\item{data_bom}{新增的BOM数据}
}
\value{
返回值
}
\description{
获取BOM新增的模板_表体部分
}
\examples{
BOM_getNewBillTpl()
}
