% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{BOM_getNewBillTpl_Head}
\alias{BOM_getNewBillTpl_Head}
\title{针对BOM的表头数据进行处理}
\usage{
BOM_getNewBillTpl_Head(
  conn = conn_vm_erp_test(),
  PMCode = "2.104.20.00034",
  PLMBatchnum = "BOM00000002",
  FInterID = 0
)
}
\arguments{
\item{conn}{连接}

\item{PMCode}{物料代码}

\item{PLMBatchnum}{批号}

\item{FInterID}{id}
}
\value{
返回值
}
\description{
针对BOM的表头数据进行处理
}
\examples{
BOM_getNewBillTpl_Head()
}
