% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_util.R
\name{BOM_getNewInterId}
\alias{BOM_getNewInterId}
\title{BOM获取最新的内码}
\usage{
BOM_getNewInterId(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
BOM获取最新的内码
}
\examples{
BOM_getNewInterId()
}
