% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bom_getVersion}
\alias{bom_getVersion}
\title{bom将PLM的版本转化为ERP中需要的版本}
\usage{
bom_getVersion(conn = conn_vm_erp_test(), version_plm = "A")
}
\arguments{
\item{conn}{连接}

\item{version_plm}{plm版本}
}
\value{
返回ERP中的版本从001开始
}
\description{
bom将PLM的版本转化为ERP中需要的版本
}
\examples{
bom_getVersion()
}
