% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D02_ERP_BOM_Model.R
\name{bom_insertValue_ERP2PLM}
\alias{bom_insertValue_ERP2PLM}
\title{insert into value}
\usage{
bom_insertValue_ERP2PLM(
  conn = conn_vm_erp_test2(),
  FNumber = "1.218.03.00002",
  FStep = 1
)
}
\arguments{
\item{conn}{conn}

\item{FNumber}{item number}

\item{FStep}{step count}
}
\value{
not value
}
\description{
insert into value
}
\examples{
bom_is_insertValue_ERP2PLM()
}
