% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D02_ERP_BOM_Model.R
\name{bom_is_haveValue_ERP2PLM}
\alias{bom_is_haveValue_ERP2PLM}
\title{initial bom to check level 1 have data and return true or}
\usage{
bom_is_haveValue_ERP2PLM(
  conn = conn_vm_erp_test2(),
  FNumber = "1.218.03.00002",
  FStep = 1
)
}
\arguments{
\item{conn}{connection}

\item{FNumber}{item number}
}
\value{
return true or false
}
\description{
initial bom to check level 1 have data and return true or
}
\examples{
bom_level1_haveValue_ERP2PLM()
}
