% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D02_ERP_BOM_Model.R
\name{bom_level_sql_ERP2PLM}
\alias{bom_level_sql_ERP2PLM}
\title{generate the sql for bom initial}
\usage{
bom_level_sql_ERP2PLM(
  conn = conn_vm_erp_test2(),
  FNumber = "1.218.03.00002",
  FStep = 1
)
}
\arguments{
\item{conn}{connection}

\item{FNumber}{the item number}

\item{FStep}{step from 1 ,2 ...}
}
\value{
return value
}
\description{
generate the sql for bom initial
}
\examples{
bom_level_sql_ERP2PLM()
}
