% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{bom_readIntoERP_updateStatus}
\alias{bom_readIntoERP_updateStatus}
\title{更新BOM的状态}
\usage{
bom_readIntoERP_updateStatus(
  conn = conn_vm_erp_test(),
  PMCode = "2.104.20.00034",
  PLMBatchnum = "BOM00000002"
)
}
\arguments{
\item{conn}{连接}

\item{PMCode}{产品代码}

\item{PLMBatchnum}{批号}
}
\value{
返回值
}
\description{
更新BOM的状态
}
\examples{
bom_readIntoERP_updateStatus()
}
