% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D02_ERP_BOM_Model.R
\name{bom_selectValue_ERP2PLM}
\alias{bom_selectValue_ERP2PLM}
\title{update the data value}
\usage{
bom_selectValue_ERP2PLM(
  conn_erp = conn_vm_erp_test2(),
  conn_plm = conn_vm_plm_prd(),
  FNumber = "1.218.03.00002"
)
}
\arguments{
\item{conn_erp}{conn1}

\item{conn_plm}{conn2}

\item{FNumber}{item number}
}
\value{
return value
}
\description{
update the data value
}
\examples{
bom_selectValue_ERP2PLM()
}
