% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_util.R
\name{BOM_setNewBillNo}
\alias{BOM_setNewBillNo}
\title{设置BOM单号的最大流水号}
\usage{
BOM_setNewBillNo(conn = conn_vm_erp_test(), ncount = 1)
}
\arguments{
\item{conn}{连接}

\item{ncount}{数量}
}
\value{
返回值
}
\description{
设置BOM单号的最大流水号
}
\examples{
BOM_setNewBillNo()
}
