% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{bom_update_bomApplyFlag}
\alias{bom_update_bomApplyFlag}
\title{处理BOM中的申请标志}
\usage{
bom_update_bomApplyFlag(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
处理BOM中的申请标志
}
\examples{
bom_update_bomApplyFlag()
}
