% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D02_ERP_BOM_Model.R
\name{bom_updateStatus_ERP2PLM}
\alias{bom_updateStatus_ERP2PLM}
\title{update the status}
\usage{
bom_updateStatus_ERP2PLM(
  conn = conn_vm_erp_test2(),
  FNumber = "1.218.03.00002"
)
}
\arguments{
\item{conn}{conn}

\item{FNumber}{the item number}
}
\value{
retur value
}
\description{
update the status
}
\examples{
bom_updateStatus_ERP2PLM()
}
