% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_conn.R
\name{conn_config_read}
\alias{conn_config_read}
\title{读取配置文件并获取连接信息}
\usage{
conn_config_read(file_name = "config/conn_erp.xlsx")
}
\arguments{
\item{file_name}{文件名}
}
\value{
返回值
}
\description{
读取配置文件并获取连接信息
}
\examples{
conn_config_read()
}
