% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_cron.R
\name{cron_set}
\alias{cron_set}
\title{同步功能设置}
\usage{
cron_set(r_file, time = 15, id = "id1", description = "Weather")
}
\arguments{
\item{r_file}{文件名}

\item{time}{时间}

\item{id}{ID}

\item{description}{描述}
}
\value{
返回值
}
\description{
同步功能设置
}
\examples{
cron_set()
}
