% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{data_PLMtoERP_BOM}
\alias{data_PLMtoERP_BOM}
\title{同步BOM信息}
\usage{
data_PLMtoERP_BOM(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接信息}
}
\value{
返回值
}
\description{
同步BOM信息
}
\examples{
data_PLMtoERP_BOM()
}
