% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{data_PLMtoERP_BOM_delInterId}
\alias{data_PLMtoERP_BOM_delInterId}
\title{增加对内码的处理}
\usage{
data_PLMtoERP_BOM_delInterId(conn = conn_vm_erp_test(), data = 1:50)
}
\arguments{
\item{conn}{连接}

\item{data}{数据}
}
\value{
返顺值
}
\description{
增加对内码的处理
}
\examples{
data_PLMtoERP_BOM_delInterId()
}
