% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{data_PLMtoERP_BOM_erp}
\alias{data_PLMtoERP_BOM_erp}
\title{获取PLM中的BOM信息}
\usage{
data_PLMtoERP_BOM_erp(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取PLM中的BOM信息
}
\examples{
data_PLMtoERP_BOM_erp()
}
