% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{data_PLMtoERP_BOM_fromDate}
\alias{data_PLMtoERP_BOM_fromDate}
\title{从PLM库中获取最新更新的数据}
\usage{
data_PLMtoERP_BOM_fromDate(
  conn_plm = conn_vm_plm_test(),
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{conn_plm}{连接}

\item{conn_erp}{ERP连接}
}
\value{
返回值
}
\description{
从PLM库中获取最新更新的数据
}
\examples{
data_PLMtoERP_Item_fromDate()
}
