% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{data_PLMtoERP_BOM_maxDate}
\alias{data_PLMtoERP_BOM_maxDate}
\title{获取最大日期}
\usage{
data_PLMtoERP_BOM_maxDate(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取最大日期
}
\examples{
data_PLMtoERP_BOM_maxDate()
}
