% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{data_PLMtoERP_Item}
\alias{data_PLMtoERP_Item}
\title{获取PLMtoERP_Item中的数据}
\usage{
data_PLMtoERP_Item(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回数据
}
\description{
获取PLMtoERP_Item中的数据
}
\examples{
data_PLMtoERP_Item()
}
