% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{data_PLMtoERP_Item_erp}
\alias{data_PLMtoERP_Item_erp}
\title{获取PLMtoERP_Item中的数据ERP数据}
\usage{
data_PLMtoERP_Item_erp(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取PLMtoERP_Item中的数据ERP数据
}
\examples{
data_PLMtoERP_Item_erp()
}
