% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{data_PLMtoERP_Item_plm}
\alias{data_PLMtoERP_Item_plm}
\title{获取PLMtoERP_Item中的数据PLM}
\usage{
data_PLMtoERP_Item_plm(conn = conn_vm_plm_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取PLMtoERP_Item中的数据PLM
}
\examples{
data_PLMtoERP_Item_plm()
}
