% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_version.R
\name{erp_bom_degrade_checkExistsInPLM}
\alias{erp_bom_degrade_checkExistsInPLM}
\title{检查在BOM在中间表中是否存在}
\usage{
erp_bom_degrade_checkExistsInPLM(
  conn_erp,
  FNumber = "2.104.20.00001",
  FBomNumber,
  FVersion_PLM = "A",
  days = 7
)
}
\arguments{
\item{conn_erp}{ERP链接}

\item{FNumber}{产品代码}

\item{FBomNumber}{产品BOM}

\item{FVersion_PLM}{BOM号}

\item{days}{最近天数}
}
\value{
返回值
}
\description{
检查在BOM在中间表中是否存在
}
\examples{
erp_bom_degrade_checkExistsInPLM()
}
