% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_version.R
\name{erp_bom_existInBom}
\alias{erp_bom_existInBom}
\title{检验BOM是否已存在}
\usage{
erp_bom_existInBom(conn_erp, FNumber = "2.104.20.00001", FVersion_PLM = "A")
}
\arguments{
\item{conn_erp}{连接信息}

\item{FNumber}{产品代码}

\item{FVersion_PLM}{产品版本}
}
\value{
返回值，存在返回TRUE,不存在返回FALSE
}
\description{
检验BOM是否已存在
}
\examples{
erp_bom_existInBom()
}
