% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_version.R
\name{erp_bom_existInLast}
\alias{erp_bom_existInLast}
\title{检查可恢复记录表中是否存在相关记录}
\usage{
erp_bom_existInLast(conn_erp, FNumber = "2.104.20.00001", FVersion_PLM = "A")
}
\arguments{
\item{conn_erp}{连接}

\item{FNumber}{产品代码}

\item{FVersion_PLM}{版本}
}
\value{
返回值
}
\description{
检查可恢复记录表中是否存在相关记录
}
\examples{
erp_bom_existInLast()
}
