% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_version.R
\name{erp_bom_existInMo}
\alias{erp_bom_existInMo}
\title{检验BOM在生产订单中是否存在，如果存在则不执行删除功能}
\usage{
erp_bom_existInMo(conn_erp, FNumber = "1.100.04.00004", FVersion_PLM = "A")
}
\arguments{
\item{conn_erp}{连接信息}

\item{FNumber}{产品代码}

\item{FVersion_PLM}{PLM版本号}
}
\value{
返回值，存在返回True,不存在返回FALSE
}
\description{
检验BOM在生产订单中是否存在，如果存在则不执行删除功能
}
\examples{
erp_bom_existInMo()
}
