% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{ERP_BOM_preCheck}
\alias{ERP_BOM_preCheck}
\title{增加BOM导入前的预检验}
\usage{
ERP_BOM_preCheck(conn = conn_vm_erp_test(), FNumber = "0.109.06.03")
}
\arguments{
\item{conn}{连接}

\item{FNumber}{代码}
}
\value{
返回值
}
\description{
增加BOM导入前的预检验
}
\examples{
ERP_BOM_preCheck()
}
