% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_version.R
\name{erp_bom_version_delete}
\alias{erp_bom_version_delete}
\title{删除ERP系统中的BOM数据，删除前正式进行表级备份，可用于后续恢复一次}
\usage{
erp_bom_version_delete(
  conn_erp,
  FNumber = "2.104.20.00001",
  FVersion_PLM = "A"
)
}
\arguments{
\item{conn_erp}{ERP连接信息}

\item{FNumber}{编码}

\item{FVersion_PLM}{版本}
}
\value{
返回值
}
\description{
删除ERP系统中的BOM数据，删除前正式进行表级备份，可用于后续恢复一次
}
\examples{
erp_bom_version_delete()
}
