% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_version.R
\name{erp_bom_version_query}
\alias{erp_bom_version_query}
\title{通过BOM返回数据}
\usage{
erp_bom_version_query(conn_erp, FNumber = "1.100.04.00004")
}
\arguments{
\item{conn_erp}{ERP连接信息}

\item{FNumber}{产品编码}
}
\value{
返回值
}
\description{
通过BOM返回数据
}
\examples{
erp_bom_version_query()
}
