% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_version.R
\name{erp_bom_version_recover}
\alias{erp_bom_version_recover}
\title{用于BOM数据的最近一次恢复}
\usage{
erp_bom_version_recover(
  conn_erp,
  FNumber = "2.104.20.00001",
  FVersion_PLM = "A"
)
}
\arguments{
\item{conn_erp}{ERPl连接}

\item{FNumber}{产品代码}

\item{FVersion_PLM}{版本}
}
\value{
返回值
}
\description{
用于BOM数据的最近一次恢复
}
\examples{
erp_bom_version_recover()
}
