% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERP_material.R
\name{erp_checkItemUseStatus}
\alias{erp_checkItemUseStatus}
\title{更新物料的状态}
\usage{
erp_checkItemUseStatus(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
更新物料的状态
}
\examples{
erp_checkItemUseStatus()
}
