% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERP_material.R
\name{erp_getItemListFromPlm}
\alias{erp_getItemListFromPlm}
\title{ERP读取PLM传入的物料数据}
\usage{
erp_getItemListFromPlm(
  conn = conn_vm_erp_test(),
  FStartDate = "2021-06-01",
  FEndDate = "2021-06-10"
)
}
\arguments{
\item{conn}{连接}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}
}
\value{
返回值
}
\description{
ERP读取PLM传入的物料数据
}
\examples{
erp_getItemListFromPlm()
}
