% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D01_ERP_Item_Model.R
\name{ERP_Item_getMaxItemNoByGroup}
\alias{ERP_Item_getMaxItemNoByGroup}
\title{根据最新的物料编码获取同分组的最大物料编码}
\usage{
ERP_Item_getMaxItemNoByGroup(
  config_file = "config/conn_k3.R",
  FNumber = "1.105.04.00001"
)
}
\arguments{
\item{config_file}{配置文件}

\item{FNumber}{物料编码}
}
\value{
返回值
}
\description{
根据最新的物料编码获取同分组的最大物料编码
}
\examples{
ERP_Item_getMaxItemNoByGroup()
}
