% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D01_ERP_Item_Model.R
\name{ERP_Item_Modify}
\alias{ERP_Item_Modify}
\title{物料修改属性}
\usage{
ERP_Item_Modify(config_file = "config/conn_k3.R", data_item)
}
\arguments{
\item{config_file}{配置文件}

\item{data_item}{物料数据}
}
\value{
返回值
}
\description{
物料修改属性
}
\examples{
ERP_Item_Modify()
}
