% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERP_Item_Model.R
\name{ERP_Item_New}
\alias{ERP_Item_New}
\title{处理物料新增数据}
\usage{
ERP_Item_New(config_file = "config/conn_k3.R", data_item)
}
\arguments{
\item{config_file}{配置文件}

\item{data_item}{物料新增}
}
\value{
返回值
}
\description{
处理物料新增数据
}
\examples{
ERP_Item_New()
}
