% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERP_material.R
\name{erp_updateItem_plmMode}
\alias{erp_updateItem_plmMode}
\title{更新物料状态}
\usage{
erp_updateItem_plmMode(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
更新物料状态
}
\examples{
erp_updateItem_plmMode()
}
