% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{item_debug_clear}
\alias{item_debug_clear}
\title{清楚BUG}
\usage{
item_debug_clear(conn = conn_vm_erp_test(), FNumber = "3.02.02.001293")
}
\arguments{
\item{conn}{链接}

\item{FNumber}{代码}
}
\value{
return value
}
\description{
清楚BUG
}
\examples{
item_debug_clear()
}
