% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{item_GetInterId}
\alias{item_GetInterId}
\title{返回物料编码的内码}
\usage{
item_GetInterId(conn_erp = conn_vm_erp_test(), FItemNumber = "0.109.06.031")
}
\arguments{
\item{conn_erp}{ERP}

\item{FItemNumber}{物料代码}
}
\value{
返回值
}
\description{
返回物料编码的内码
}
\examples{
item_GetInterId()
}
