% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_getItemID_costObj}
\alias{Item_getItemID_costObj}
\title{to get the obj Itemid}
\usage{
Item_getItemID_costObj(conn = conn_vm_erp_test2(), FNumber = "RDS.02.000002")
}
\arguments{
\item{conn}{conn}

\item{FNumber}{new number for prebook}
}
\value{
return value
}
\description{
to get the obj Itemid
}
\examples{
Item_getItemID_costObj()
}
