% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_getList}
\alias{Item_getList}
\title{获取物料的处理清单}
\usage{
Item_getList(conn = conn_vm_erp_test())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取物料的处理清单
}
\examples{
Item_getList()
}
