% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Item.R
\name{item_getNumber}
\alias{item_getNumber}
\title{获取物料编码}
\usage{
item_getNumber(conn_erp = conn_vm_erp_test(), item_list)
}
\arguments{
\item{conn_erp}{ERP连接信息}

\item{item_list}{列表}
}
\value{
返回值
}
\description{
获取物料编码
}
\examples{
item_getNumber()
}
