% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_GetParentItemId_costObj}
\alias{Item_GetParentItemId_costObj}
\title{获取成本对象的代码}
\usage{
Item_GetParentItemId_costObj(
  conn = conn_vm_erp_test(),
  MCode = "3.02.09.001025"
)
}
\arguments{
\item{conn}{链接}

\item{MCode}{物料}
}
\value{
返回值
}
\description{
获取成本对象的代码
}
\examples{
Item_GetParentItemId_costObj()
}
