% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_getUnAllocateItemId}
\alias{Item_getUnAllocateItemId}
\title{获取未分配的物料内码}
\usage{
Item_getUnAllocateItemId(conn = conn_vm_erp_test(), FNumber = "")
}
\arguments{
\item{conn}{连接}

\item{FNumber}{物料代码}
}
\value{
返回值
}
\description{
获取未分配的物料内码
}
\examples{
Item_getUnAllocateItemId()
}
