% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_getUnAllocateNumber}
\alias{Item_getUnAllocateNumber}
\title{返回未分配的物料数据}
\usage{
Item_getUnAllocateNumber(conn = conn_vm_erp_test(), MProp = "外购")
}
\arguments{
\item{conn}{连接}

\item{MProp}{物料属性}

\item{n}{返回数量}
}
\value{
返回值
}
\description{
返回未分配的物料数据
}
\examples{
Item_getUnAllocateNumber()
}
