% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_Initial_WG}
\alias{Item_Initial_WG}
\title{写入外购的物料进行待分配表}
\usage{
Item_Initial_WG(
  conn = conn_vm_erp_test(),
  table_name_rds = "t_Item_rds",
  table_name_room = "t_item_rdsRoom"
)
}
\arguments{
\item{conn}{连接}

\item{table_name_rds}{外购物料}

\item{table_name_room}{待分配表}
}
\value{
返回值
}
\description{
写入外购的物料进行待分配表
}
\examples{
Item_Initial_WG()
}
