% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_Initial_WW}
\alias{Item_Initial_WW}
\title{处理委外待分配数据}
\usage{
Item_Initial_WW(
  conn = vmrdspkg::conn_vm_erp_test(),
  table_name_rds = "t_Item_rds",
  table_name_room = "t_item_rdsRoom"
)
}
\arguments{
\item{conn}{连接}

\item{table_name_rds}{物料明细表}

\item{table_name_room}{待分配信息表}
}
\value{
返回值
}
\description{
处理委外待分配数据
}
\examples{
md_pushItem2UnAllocated_ww()
}
