% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_Initial_ZZ}
\alias{Item_Initial_ZZ}
\title{针对自制物料进行初始化隐藏处理}
\usage{
Item_Initial_ZZ(
  conn = conn_vm_erp_test(),
  table_name_rds = "t_Item_rds",
  table_name_room = "t_item_rdsRoom"
)
}
\arguments{
\item{conn}{ERP链接信息}

\item{table_name_rds}{物料明细表}

\item{table_name_room}{待分配统计表}
}
\value{
返回值
}
\description{
针对自制物料进行初始化隐藏处理
}
\examples{
Item_Initial_ZZ()
}
