% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_readIntoERP_One}
\alias{Item_readIntoERP_One}
\title{写入一行物料数据}
\usage{
Item_readIntoERP_One(conn = conn_vm_erp_test(), MCode, MProp, PLMBatchnum)
}
\arguments{
\item{conn}{连接}

\item{MCode}{物料编码}

\item{MProp}{属性}

\item{PLMBatchnum}{批次号}
}
\value{
返回值
}
\description{
写入一行物料数据
}
\examples{
Item_readIntoERP_One()
}
