% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_ReadItem_One}
\alias{Item_ReadItem_One}
\title{读取数据}
\usage{
Item_ReadItem_One(
  conn = conn_vm_erp_test(),
  MCode = "3.02.09.001025",
  MProp = "外购",
  PLMBatchnum = "ECN00000002"
)
}
\arguments{
\item{conn}{连接}

\item{MCode}{物料}

\item{MProp}{属性}

\item{PLMBatchnum}{批次}
}
\value{
返回值
}
\description{
读取数据
}
\examples{
Item_ReadItem_One
}
