% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D01_ERP_Item_Model.R
\name{item_selectValue_ERP2PLM_multi}
\alias{item_selectValue_ERP2PLM_multi}
\title{get the initial item value}
\usage{
item_selectValue_ERP2PLM_multi(
  conn_erp = conn_vm_erp_test2(),
  conn_plm = conn_vm_plm_prd(),
  FNumbers
)
}
\arguments{
\item{conn_erp}{conn1}

\item{conn_plm}{conn2}

\item{FNumbers}{增加多选的物料编码}
}
\value{
return value
}
\description{
get the initial item value
}
\examples{
item_selectValue_ERP2PLM_multi()
}
