% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D01_ERP_Item_Model.R
\name{item_selectValue_ERP2PLM_newWrite}
\alias{item_selectValue_ERP2PLM_newWrite}
\title{填写相应的初始化物料数据}
\usage{
item_selectValue_ERP2PLM_newWrite(
  conn_erp = conn_vm_erp_test2(),
  conn_plm = conn_vm_plm_prd()
)
}
\arguments{
\item{conn_erp}{ERP连接}

\item{conn_plm}{PLM连接}
}
\value{
返回值
}
\description{
填写相应的初始化物料数据
}
\examples{
item_selectValue_ERP2PLM_newWrite()
}
