% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item2.R
\name{Item_updateTaskStatus_One}
\alias{Item_updateTaskStatus_One}
\title{更新任务单状态}
\usage{
Item_updateTaskStatus_One(
  conn = conn_vm_erp_test(),
  MCode = "3.02.09.001025",
  MProp = "外购",
  PLMBatchnum = "ECN00000002"
)
}
\arguments{
\item{conn}{连接}

\item{MCode}{物料代码}

\item{MProp}{物料名称}

\item{PLMBatchnum}{批次号}
}
\value{
返回值
}
\description{
更新任务单状态
}
\examples{
Item_updateTaskStatus_One()
}
