% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E01_log.R
\name{log_detail}
\alias{log_detail}
\title{日志明细信息}
\usage{
log_detail(
  conn = conn_vm_plm_test(),
  FTableName = "rds_vw_log_PLMtoERP_Item_detail",
  FLogDate = "2021-04-22"
)
}
\arguments{
\item{conn}{连接}

\item{FTableName}{表名}

\item{FLogDate}{日志日期}
}
\value{
返回值
}
\description{
日志明细信息
}
\examples{
log_detail()
}
