% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E01_log.R
\name{log_summary}
\alias{log_summary}
\title{日志汇总信息}
\usage{
log_summary(
  conn = conn_vm_plm_test(),
  FTableName = "rds_vw_log_PLMtoERP_Item_summary",
  FStartDate = "2021-01-01",
  FEndDate = "2021-08-01"
)
}
\arguments{
\item{conn}{连接}

\item{FTableName}{表名}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}
}
\value{
返回值
}
\description{
日志汇总信息
}
\examples{
log_summary()
}
