% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Item.R
\name{PLM2ERP_Item_getBatchNo}
\alias{PLM2ERP_Item_getBatchNo}
\title{读取物料的批次列表}
\usage{
PLM2ERP_Item_getBatchNo(config_file = "config/conn_tc.R")
}
\arguments{
\item{config_file}{配置文件}
}
\value{
返回值
}
\description{
读取物料的批次列表
}
\examples{
PLM2ERP_Item_getBatchNo()
}
