% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D04_PLM2ERP_Bom.R
\name{PLM_BOM_ECN}
\alias{PLM_BOM_ECN}
\title{PLM-BOM的工程变更与版本升级}
\usage{
PLM_BOM_ECN(config_file = "config/conn_k3.R", data_bom)
}
\arguments{
\item{config_file}{配置文件}

\item{data_bom}{BOM数据}
}
\value{
返回值
}
\description{
PLM-BOM的工程变更与版本升级
}
\examples{
PLM_BOM_ECN()
}
