% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item.R
\name{PLM_Item_Allocated_wg}
\alias{PLM_Item_Allocated_wg}
\title{针对外购物料进行分配}
\usage{
PLM_Item_Allocated_wg(
  config_file = "config/conn_tc.R",
  batchNo = "APP00000005",
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{config_file}{配置文件}

\item{batchNo}{批号}

\item{conn_erp}{ERP连接信息}
}
\value{
返回值
}
\description{
针对外购物料进行分配
}
\examples{
PLM_Item_Allocated_wg()
}
