% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Item.R
\name{PLM_Item_Allocated_ww}
\alias{PLM_Item_Allocated_ww}
\title{委外加工}
\usage{
PLM_Item_Allocated_ww(
  config_file = "config/conn_tc.R",
  batchNo = "APP00000005",
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{config_file}{连接}

\item{batchNo}{批号}

\item{conn_erp}{连接}
}
\value{
返回值
}
\description{
委外加工
}
\examples{
PLM_Item_Allocated_ww()
}
