% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Item.R
\name{PLM_Item_Allocated_zz}
\alias{PLM_Item_Allocated_zz}
\title{自制物料的分配-----}
\usage{
PLM_Item_Allocated_zz(
  config_file = "config/conn_tc.R",
  batchNo = "APP00000005",
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{config_file}{配置文件}

\item{batchNo}{批次}

\item{conn_erp}{物料}
}
\value{
返回值
}
\description{
自制物料的分配-----
}
\examples{
PLM_Item_Allocated_zz()
}
