% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item.R
\name{PLM_Item_App_getBatchNo}
\alias{PLM_Item_App_getBatchNo}
\title{读取物料申请的批次列表}
\usage{
PLM_Item_App_getBatchNo(config_file = "config/conn_tc.R")
}
\arguments{
\item{config_file}{配置文件}
}
\value{
返回值
}
\description{
读取物料申请的批次列表
}
\examples{
PLM_Item_App_getBatchNo()
}
