% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Item.R
\name{PLM_Item_getBatchNo_Aux}
\alias{PLM_Item_getBatchNo_Aux}
\title{读取物料申请的批次列表}
\usage{
PLM_Item_getBatchNo_Aux(config_file = "config/conn_tc.R", prefix = "APP")
}
\arguments{
\item{config_file}{配置文件}

\item{prefix}{批号的前缀}
}
\value{
返回值
}
\description{
读取物料申请的批次列表
}
\examples{
PLM_Item_getBatchNo_Aux()
}
