% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Item.R
\name{PLM_Item_readByBatchNo}
\alias{PLM_Item_readByBatchNo}
\title{按批次号获取新增物料信息}
\usage{
PLM_Item_readByBatchNo(
  config_file = "config/conn_tc.R",
  batchNo = "APP00000005"
)
}
\arguments{
\item{config_file}{配置文件}

\item{batchNo}{批号}
}
\value{
返回值
}
\description{
按批次号获取新增物料信息
}
\examples{
PLM2ERP_Item_readByBatchNo()
}
