% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Item.R
\name{PLM_Item_readByBatchNo_WG}
\alias{PLM_Item_readByBatchNo_WG}
\title{外购物料的读取}
\usage{
PLM_Item_readByBatchNo_WG(
  config_file = "config/conn_tc.R",
  batchNo = "APP00000005"
)
}
\arguments{
\item{config_file}{配置文件}

\item{batchNo}{批次号}
}
\value{
返回所有相关的物料
}
\description{
外购物料的读取
}
\examples{
PLM_Item_readByBatchNo_WG()
}
