% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D03_PLM2ERP_Item.R
\name{PLM_Item_readByBatchNo_WG_New}
\alias{PLM_Item_readByBatchNo_WG_New}
\title{获取全新的外购物料}
\usage{
PLM_Item_readByBatchNo_WG_New(
  config_file = "config/conn_tc.R",
  batchNo = "APP00000005",
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{config_file}{连接配置文件}

\item{batchNo}{批次号}

\item{conn_erp}{ERP连接信息}
}
\value{
返回值
}
\description{
获取全新的外购物料
}
\examples{
PLM_Item_readByBatchNo_WG_New()
}
