% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Item.R
\name{PLM_Item_readByBatchNo_WW_New}
\alias{PLM_Item_readByBatchNo_WW_New}
\title{获取最新的委外物料信息}
\usage{
PLM_Item_readByBatchNo_WW_New(
  config_file = "config/conn_tc.R",
  batchNo = "APP00000005",
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{config_file}{配置文件}

\item{batchNo}{批号}

\item{conn_erp}{ERP连接信息}
}
\value{
返回值
}
\description{
获取最新的委外物料信息
}
\examples{
PLM_Item_readByBatchNo_WW_New()
}
