% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_bom.R
\name{rpt_icmo_ppbom_diff_Item}
\alias{rpt_icmo_ppbom_diff_Item}
\title{按生产任务单查询生产投料及BOM的差异}
\usage{
rpt_icmo_ppbom_diff_Item(conn = conn_vm_erp_test(), fnumber = "1.207.01.00002")
}
\arguments{
\item{conn}{连接}

\item{fbillno}{单据编码}
}
\value{
返回值
}
\description{
按生产任务单查询生产投料及BOM的差异
}
\examples{
rpt_icmo_ppbom_diff_Item()
}
