% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_util.R
\name{sql_Fnumber_multi}
\alias{sql_Fnumber_multi}
\title{针对物料编码进行处理}
\usage{
sql_Fnumber_multi(FNumbers = "100.01,100.02")
}
\arguments{
\item{FNumbers}{物料编码}
}
\value{
返回值
}
\description{
针对物料编码进行处理
}
\examples{
sql_Fnumber_multi()
}
