% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{sync_PLMtoERP_BOM_intially}
\alias{sync_PLMtoERP_BOM_intially}
\title{同步BOM信息到ERP库}
\usage{
sync_PLMtoERP_BOM_intially(
  conn_plm = conn_vm_plm_test(),
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{conn_plm}{PLM连接信息}

\item{conn_erp}{ERP连接信息}
}
\value{
返回值
}
\description{
同步BOM信息到ERP库
}
\examples{
sync_PLMtoERP_BOM_intially()
表名：
SQL名：
}
