% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{sync_PLMtoERP_BOM_periodly}
\alias{sync_PLMtoERP_BOM_periodly}
\title{周期性同步BOM数据}
\usage{
sync_PLMtoERP_BOM_periodly(
  conn_plm = conn_vm_plm_test(),
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{conn_plm}{PLM连接}

\item{conn_erp}{ERP连接}
}
\value{
返回值
}
\description{
周期性同步BOM数据
}
\examples{
sync_PLMtoERP_BOM_periodly()
}
