% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sync.R
\name{sync_PLMtoERP_Item_initially}
\alias{sync_PLMtoERP_Item_initially}
\title{同步物料信息到ERP库,初始化仅一次}
\usage{
sync_PLMtoERP_Item_initially(
  conn_plm = conn_vm_plm_test(),
  conn_erp = conn_vm_erp_test()
)
}
\arguments{
\item{conn_plm}{PLM连接信息}

\item{conn_erp}{ERP连接信息}
}
\value{
返回值
}
\description{
同步物料信息到ERP库,初始化仅一次
}
\examples{
sync_PLMtoERP_Item()
表名：
SQL名：

}
