% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLM2ERP_Bom.R
\name{BOM_getNewBillTpl_Body}
\alias{BOM_getNewBillTpl_Body}
\title{获取BOM新增的模板_表体部分}
\usage{
BOM_getNewBillTpl_Body(
  conn = conn_vm_erp_test(),
  BOMRevCode = "TCB100181/A",
  PLMBatchnum = "BOM00000002",
  FLowCode = 2
)
}
\arguments{
\item{conn}{连接}

\item{BOMRevCode}{子项BOM版本号}

\item{PLMBatchnum}{批号}

\item{FLowCode}{流程低位码}
}
\value{
返回值
}
\description{
获取BOM新增的模板_表体部分
}
\examples{
BOM_getNewBillTpl_Body()
}
