% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERP_material.R
\name{erp_materia_read}
\alias{erp_materia_read}
\title{物料属性批量修改}
\usage{
erp_materia_read(
  file = "data-raw/VM物料批量修改模板.xlsx",
  sheet = "物料",
  lang = "cn"
)
}
\arguments{
\item{file}{文件名}

\item{sheet}{页答}

\item{lang}{语言}
}
\value{
返回值
}
\description{
物料属性批量修改
}
\examples{
erp_materia_read()
}
